+++ packages/atproto_core/lib/atproto_core.dart
@@ -6,6 +6,7 @@ export 'package:at_uri/at_uri.dart' show AtUri;
 export 'package:atproto_core/src/base_service.dart';
 export 'package:atproto_core/src/car_decoder.dart';
 export 'package:atproto_core/src/platform.dart';
+export 'package:atproto_core/src/utils.dart' show isValidAppPassword;
 export 'package:atproto_core/src/clients/challenge.dart';
 export 'package:atproto_core/src/paginations/pagination.dart';
 export 'package:atproto_core/src/clients/client_context.dart';