+++ packages/atproto/lib/atproto.dart
@@ -35,7 +35,6 @@ export 'package:atproto/src/entities/repo.dart';
 export 'package:atproto/src/entities/repo_block.dart';
 export 'package:atproto/src/entities/repo_blocks.dart';
 export 'package:atproto/src/entities/repo_commit.dart';
-export 'package:atproto/src/entities/repo_commit_paths.dart';
 export 'package:atproto/src/entities/repo_commits.dart';
 export 'package:atproto/src/entities/repo_head.dart';
 export 'package:atproto/src/entities/repo_info.dart';