+++ packages/atproto/test/src/sync/sync_service_test.dart
@@ -10,7 +10,6 @@ import 'package:test/test.dart';
 // 🌎 Project imports:
 import 'package:atproto/src/entities/repo_blocks.dart';
 import 'package:atproto/src/entities/repo_commit.dart';
-import 'package:atproto/src/entities/repo_commit_paths.dart';
 import 'package:atproto/src/entities/repo_commits.dart';
 import 'package:atproto/src/entities/repo_head.dart';
 import 'package:atproto/src/entities/repos.dart';
@@ -109,100 +108,6 @@ void main() {
 
       atp_test.expectRateLimitExceededException(
         () async => await sync.findRepoCommits(
-          did: 'did:plc:jb3pkzwuhnmq65ktmib27eli',
-        ),
-      );
-    });
-  });
-
-  group('.findRepoCommitPaths', () {
-    test('normal case', () async {
-      final sync = SyncService(
-        did: 'test',
-        protocol: core.Protocol.https,
-        service: 'test',
-        context: core.ClientContext(
-          accessJwt: '1234',
-          timeout: Duration.zero,
-        ),
-        mockedGetClient: atp_test.createMockedGetClient(
-          'test/src/sync/data/find_repo_commit_paths.json',
-        ),
-      );
-
-      final response = await sync.findRepoCommitPaths(
-        did: 'did:plc:jb3pkzwuhnmq65ktmib27eli',
-        earliestCommitCid: 'baaaaaaa',
-        latestCommitCid: 'baaaaaaa',
-      );
-
-      expect(response, isA<core.XRPCResponse>());
-      expect(response.data, isA<RepoCommitPaths>());
-    });
-
-    test('as JSON', () async {
-      final sync = SyncService(
-        did: 'test',
-        protocol: core.Protocol.https,
-        service: 'test',
-        context: core.ClientContext(
-          accessJwt: '1234',
-          timeout: Duration.zero,
-        ),
-        mockedGetClient: atp_test.createMockedGetClient(
-          'test/src/sync/data/find_repo_commit_paths.json',
-        ),
-      );
-
-      final response = await sync.findRepoCommitPathsAsJson(
-        did: 'did:plc:jb3pkzwuhnmq65ktmib27eli',
-        earliestCommitCid: 'baaaaaaa',
-        latestCommitCid: 'baaaaaaa',
-      );
-
-      expect(response, isA<core.XRPCResponse>());
-      expect(response.data, isA<Map<String, dynamic>>());
-    });
-
-    test('when unauthorized', () async {
-      final sync = SyncService(
-        did: 'test',
-        protocol: core.Protocol.https,
-        service: 'test',
-        context: core.ClientContext(
-          accessJwt: '1234',
-          timeout: Duration.zero,
-        ),
-        mockedGetClient: atp_test.createMockedGetClient(
-          'test/src/data/error.json',
-          statusCode: 401,
-        ),
-      );
-
-      atp_test.expectUnauthorizedException(
-        () async => await sync.findRepoCommitPaths(
-          did: 'did:plc:jb3pkzwuhnmq65ktmib27eli',
-        ),
-      );
-    });
-
-    test('when rate limit exceeded', () async {
-      final sync = SyncService(
-        did: 'test',
-        protocol: core.Protocol.https,
-        service: 'test',
-        context: core.ClientContext(
-          accessJwt: '1234',
-          timeout: Duration.zero,
-        ),
-        mockedGetClient: atp_test.createMockedGetClient(
-          'test/src/data/error.json',
-          statusCode: 429,
-        ),
-      );
-
-      atp_test.expectRateLimitExceededException(
-        () async => await sync.findRepoCommitPaths(
           did: 'did:plc:jb3pkzwuhnmq65ktmib27eli',
         ),
       );