+++ packages/bluesky/lib/bluesky.dart
@@ -22,8 +22,6 @@ export 'package:atproto/atproto.dart'
         $CurrentSessionCopyWith,
         Account,
         $AccountCopyWith,
-        RepoCommitPaths,
-        $RepoCommitPathsCopyWith,
         CreatedInviteCode,
         $CreatedInviteCodeCopyWith,
         CreatedInviteCodes,