+++ packages/bluesky/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Supported `app.bsky.notification.registerPush` as `createPushRegistration` on `NotificationsService`. ([#709](https://github.com/myConsciousness/atproto.dart/issues/709))
 - Added `isValidAppPassword` function. ([#713](https://github.com/myConsciousness/atproto.dart/issues/713))
+- Removed fully deprecated endpoint on `SyncService`. ([#716](https://github.com/myConsciousness/atproto.dart/issues/716))
+  - `findRepoCommitPaths`
+  - `findRepoCommitPathsAsJson`
 
 ## v0.8.4